// src/lib/api/generated/models/referenceAttributeInputDtoMatchType.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type ReferenceAttributeInputDtoMatchType = typeof ReferenceAttributeInputDtoMatchType[keyof typeof ReferenceAttributeInputDtoMatchType];


export const ReferenceAttributeInputDtoMatchType = {
  COMPATIBLE: 'COMPATIBLE',
  NOT_COMPATIBLE: 'NOT_COMPATIBLE',
  BOOST: 'BOOST',
} as const;